Create Portfolio-Page
Many players dedicate a pet page to keep up with their accomplishments and goals on the website as it helps players connect in the community. This was created by pulling elements from 4 different premade open source templates and combining them into what I wanted. Links to all original codes can be found in the credits section.