Create Neopet-Description
This is the main lookup of a specific pet. Once again I used a premade open source code and made modifications to change the background. In order to add the link to the background source I had to reposition the link so it did not overlap with the CSS source link.